Reorganized app-related code within the package
