+++ src/remarx/__init__.py
@@ -3,8 +3,8 @@
 # Copyright (c) 2025, Center for Digital Humanities, Princeton University
 # SPDX-License-Identifier: Apache-2.0
 
-from remarx import app, app_utils, sentence
+from remarx import app, sentence
 
 __version__ = "0.2.0"
 
-__all__ = ["__version__", "app", "app_utils", "sentence"]
+__all__ = ["__version__", "app", "sentence"]
