+++ src/remarx/app/__init__.py
@@ -1,0 +1,7 @@
+"""
+This module contains the UI and utilities for the remarx app
+"""
+
+from remarx.app import ui, utils
+
+__all__ = ["ui", "utils"]
