+++ src/remarx/app/ui.py
@@ -23,7 +23,7 @@ def _():
     import tempfile
 
     import remarx
-    from remarx.app_utils import create_temp_input
+    from remarx.app.utils import create_temp_input
     from remarx.sentence.corpus.create import create_corpus
     from remarx.sentence.corpus import FileInput
     return FileInput, create_corpus, create_temp_input, mo, pathlib, remarx
